requirements.txt

numpy
pillow
opencv-python
pytesseract
